Sixteen note buses, and a block that arrives on none of them
Two things, and the second is why the first was worth having.

**N_NOTE_BUS goes from 8 to 16.** Eight was the count everything else in
the machine happened to be, and a patch that splits a keyboard three ways
and sends each zone somewhere ran out of them. The masks were already
16-bit and the static_asserts already said so, so the only thing in the
way was the scheduler: it holds the three domains in one index space, and
16 + 8 + 8 was exactly the 32 bits it had. That word is 64 bits now, which
leaves room the other two domains do not have to ask for. On the board it
is 2 KB of note queues and the link has 236 KB of RAM1 to spare.

**And a new block no longer lands on a bus somebody else is on.** The
editor used to patch a node in on arrival - required inlets onto the bus
the last node wrote, the first outlet onto the first free bus - and a
MIDI port onto whichever bus was waiting. Every one of those was a guess
about a patch that did not exist yet, and on a canvas a guess is an arrow
nobody drew: add an output while a chain is being built and the chain's
next node feeds it; add two unrelated nodes and they come out joined.
Undoing one means working out which port to move off which bus.

So `connectNewNode` is gone. A node is added with every port at NO_BUS
and the wires are the ones you drag. A jack and a MIDI port carry a bus
index whatever they are doing, so "unconnected" for one of them is a bus
nothing else is on - `unusedBus`, which refuses rather than merging when
the domain is full.

The other half of that rule is what the firmware already enforced: it
refuses a node whose required inlet is empty, so the app has to refuse to
send one, or every incremental edit afterwards addresses a node the module
never took. `validate` says so and `sendWhole` acts on it, both already
there; the protocol suite now checks the two agree with the firmware for
every algorithm the module has.

+++ src/node/schedule.cpp
@@ -5,11 +5,11 @@
 static const uint8_t NOTE_BASE = N_GATE_BUS;
 static const uint8_t CV_BASE = N_GATE_BUS + N_NOTE_BUS;
 
-static uint32_t bus_bit(Domain domain, uint8_t bus){
+static uint64_t bus_bit(Domain domain, uint8_t bus){
     switch (domain){
-        case Domain::Gate: return bus < N_GATE_BUS ? (uint32_t)1u << bus : 0u;
-        case Domain::Note: return bus < N_NOTE_BUS ? (uint32_t)1u << (NOTE_BASE + bus) : 0u;
-        case Domain::CV:   return bus < N_CV_BUS   ? (uint32_t)1u << (CV_BASE + bus) : 0u;
+        case Domain::Gate: return bus < N_GATE_BUS ? (uint64_t)1u << bus : 0u;
+        case Domain::Note: return bus < N_NOTE_BUS ? (uint64_t)1u << (NOTE_BASE + bus) : 0u;
+        case Domain::CV:   return bus < N_CV_BUS   ? (uint64_t)1u << (CV_BASE + bus) : 0u;
     }
     return 0u;
 }
@@ -66,7 +66,7 @@ void Schedule::build(){
     uint64_t writers[N_BUS_TOTAL] = {0};
     for (uint8_t i = 0; i < n; i++){
         for (uint8_t b = 0; b < N_BUS_TOTAL; b++){
-            if (ports[i].write & ((uint32_t)1u << b)) writers[b] |= (uint64_t)1u << i;
+            if (ports[i].write & ((uint64_t)1u << b)) writers[b] |= (uint64_t)1u << i;
         }
     }
     // The key, the same way: whoever writes it runs before whoever plays in
@@ -79,7 +79,7 @@ void Schedule::build(){
     for (uint8_t i = 0; i < n; i++){
         uint64_t before_this = 0;
         for (uint8_t b = 0; b < N_BUS_TOTAL; b++){
-            if (ports[i].read & ((uint32_t)1u << b)) before_this |= writers[b];
+            if (ports[i].read & ((uint64_t)1u << b)) before_this |= writers[b];
         }
         if (ports[i].reads_key) before_this |= key_writers;
         // A node reading a bus it writes itself is a loop of one: it reads
@@ -116,7 +116,7 @@ void Schedule::build(){
     for (uint8_t b = 0; b < N_BUS_TOTAL; b++){
         uint8_t last = 0xFF;
         for (uint8_t pos = 0; pos < n; pos++){
-            if (ports[order[pos]].write & ((uint32_t)1u << b)) last = pos;
+            if (ports[order[pos]].write & ((uint64_t)1u << b)) last = pos;
         }
         add_bus(last == 0xFF ? early : publish[last], b);
     }
